Factor out e2e Options type to e2e/util package
This commit only tracks moving the code and the needed cleanup
and code refactoring (e.g. exporting, fixing imports etc) will
be tracked in the next commit.
